Start work on a headless QPA plugin to allow Qt to be used without an X server in linux.
Useful for the various command line tools that require
Qt. For example conversion with PDF Output or MOBI output +
rasterization of SVG images. Or ebook-meta with rendering of first page
of EPUB as cover. Or downloading news using a WebKit browser as the
backend.

Makes it unnecessary to use xvfb and will hopefully end the pointless
bug reports from people that try to use calibre command line tools in
environments without X.